+++ app/src/main/res/layout/activity_password.xml
@@ -31,11 +31,12 @@
     <ProgressBar
         android:id="@+id/passwordProgressBar"
         style="?android:attr/progressBarStyle"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="4dp"
+        android:layout_width="140dp"
+        android:layout_height="182dp"
+        android:layout_marginTop="312dp"
         android:visibility="gone"
         app:layout_constraintEnd_toEndOf="@+id/resetPasswordButton"
+        app:layout_constraintHorizontal_bias="0.0"
         app:layout_constraintStart_toStartOf="@+id/resetPasswordButton"
         app:layout_constraintTop_toBottomOf="@+id/editTextEmailPassword" />
 </androidx.constraintlayout.widget.ConstraintLayout>
